requirements.txt
streamlit 
anthropic 
requests 
python-dotenv